fix: Resolve audiobook compilation errors
- Fix OfflineAudioPlayer seekTo signature (Long -> Int)
- Fix PlaybackControls onProgress callback syntax
- Comment out missing AI imports
- Replace AI summary with text fallback

+++ app/src/main/java/com/secondmind/minimal/utils/OfflineAudioPlayer.kt
@@ -46,7 +46,7 @@ class OfflineAudioPlayer {
         }
     }
 
-    fun seekTo(position: Long) {
+    fun seekTo(position: Int) {
         player?.seekTo(position.toInt())
     }
